refactor: Rename ObjectFilter to PropertyPathFilter
