+++ library/ExternalContent/PropertyPathFilter/Contracts/Enums/Relation.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Municipio\ExternalContent\PropertyPathFilter\Contracts\Enums;
+
+enum Relation
+{
+    case AND;
+    case OR;
+}
