+++ library/ExternalContent/PropertyPathFilter/Contracts/RuleInterface.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Municipio\ExternalContent\ObjectFilter\Contracts;
+namespace Municipio\ExternalContent\PropertyPathFilter\Contracts;
 
-use Municipio\ExternalContent\ObjectFilter\Contracts\Enums\Operator;
-use Municipio\ExternalContent\ObjectFilter\Contracts\Enums\Relation;
+use Municipio\ExternalContent\PropertyPathFilter\Contracts\Enums\Operator;
+use Municipio\ExternalContent\PropertyPathFilter\Contracts\Enums\Relation;
 
 interface RuleInterface
 {
