+++ library/ExternalContent/PropertyPathFilter/Filter.test.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Municipio\ExternalContent\ObjectFilter;
+namespace Municipio\ExternalContent\PropertyPathFilter;
 
-use Municipio\ExternalContent\ObjectFilter\Contracts\RuleSetInterface;
+use Municipio\ExternalContent\PropertyPathFilter\Contracts\RuleSetInterface;
 use PHPUnit\Framework\TestCase;
 
 class FilterTest extends TestCase
