+++ library/ExternalContent/PropertyPathFilter/Rule.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Municipio\ExternalContent\ObjectFilter;
+namespace Municipio\ExternalContent\PropertyPathFilter;
 
-use Municipio\ExternalContent\ObjectFilter\Contracts\Enums\Operator;
-use Municipio\ExternalContent\ObjectFilter\Contracts\Enums\Relation;
-use Municipio\ExternalContent\ObjectFilter\Contracts\RuleInterface;
+use Municipio\ExternalContent\PropertyPathFilter\Contracts\Enums\Operator;
+use Municipio\ExternalContent\PropertyPathFilter\Contracts\Enums\Relation;
+use Municipio\ExternalContent\PropertyPathFilter\Contracts\RuleInterface;
 
 /**
  * Rule class.
