+++ library/ExternalContent/PropertyPathFilter/RuleSet.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Municipio\ExternalContent\ObjectFilter;
+namespace Municipio\ExternalContent\PropertyPathFilter;
 
-use Municipio\ExternalContent\ObjectFilter\Contracts\Enums\Relation;
-use Municipio\ExternalContent\ObjectFilter\Contracts\RuleSetInterface;
-use Municipio\ExternalContent\ObjectFilter\Contracts\RuleInterface;
+use Municipio\ExternalContent\PropertyPathFilter\Contracts\Enums\Relation;
+use Municipio\ExternalContent\PropertyPathFilter\Contracts\RuleSetInterface;
+use Municipio\ExternalContent\PropertyPathFilter\Contracts\RuleInterface;
 
 /**
  * RuleSet class.
